Move and rename document-test.js file
